refactor: extract memos CLI runner into separate package and update tests
