+++ cmd/cli/memos/runner/helpers_test.go
@@ -1,4 +1,4 @@
-package main
+package runner
 
 import (
 	"bytes"
