+++ cmd/cli/memos/runner/run_parse_error_test.go
@@ -1,4 +1,4 @@
-package main
+package runner
 
 import (
 	"bytes"
@@ -13,7 +13,7 @@ func TestRun_ParseError_Error(t *testing.T) {
 	var stdout bytes.Buffer
 	var stderr bytes.Buffer
 
-	exitCode := run([]string{
+	exitCode := Run([]string{
 		"-operation=get-memo",
 		"-base-url=https://memos.example.com",
 	}, &stdout, &stderr, func(conf *cfg.Config) usecases.MemoService {
